Moving loading functions to table classes.
