+++ src/teehr/evaluation/evaluation.py
@@ -10,14 +10,9 @@ import logging
 from teehr.loading.utils import (
     copy_template_to,
 )
-from teehr.loading.joined_timeseries import (
-    create_joined_timeseries_dataset,
-)
 import teehr.const as const
 from teehr.evaluation.fetch import Fetch
-from teehr.evaluation.load import Load
 from teehr.evaluation.metrics import Metrics
-# from teehr.evaluation.fields import Fields
 from teehr.evaluation.tables import (
     UnitTable,
     VariableTable,
@@ -104,11 +99,6 @@ class Evaluation:
     def fetch(self) -> Fetch:
         """The fetch component class."""
         return Fetch(self)
-
-    @property
-    def load(self) -> Load:
-        """The load component class."""
-        return Load(self)
 
     @property
     def metrics(self) -> Metrics:
@@ -204,18 +194,3 @@ class Evaluation:
         Includes retrieving metadata and contents.
         """
         pass
-
-    def create_joined_timeseries(self, execute_udf: bool = False):
-        """Create joined timeseries.
-
-        Parameters
-        ----------
-        execute_udf : bool, optional
-            Execute UDFs, by default False
-        """
-        create_joined_timeseries_dataset(
-            self.spark,
-            self.dataset_dir,
-            self.scripts_dir,
-            execute_udf,
-        )
