+++ src/teehr/evaluation/tables.py
@@ -1,4 +1,6 @@
 """Access methods to units table."""
+from pathlib import Path
+
 from teehr.models.str_enum import StrEnum
 from teehr.querying.filter_format import validate_and_apply_filters
 import pyspark.sql as ps
@@ -38,6 +40,32 @@ from teehr.models.table_enums import (
     TimeseriesFields,
     JoinedTimeseriesFields
 )
+from teehr.loading.timeseries import (
+    convert_timeseries,
+    validate_and_insert_timeseries
+)
+from teehr.loading.add_domains import (
+    add_configuration,
+    add_unit,
+    add_variable,
+    add_attribute
+)
+from teehr.loading.locations import (
+    convert_locations,
+    validate_and_insert_locations
+)
+from teehr.loading.location_crosswalks import (
+    convert_location_crosswalks,
+    validate_and_insert_location_crosswalks
+)
+from teehr.loading.location_attributes import (
+    convert_location_attributes,
+    validate_and_insert_location_attributes
+)
+from teehr.loading.joined_timeseries import (
+    create_joined_timeseries_dataset
+)
+import teehr.const as const
 
 import logging
 
@@ -56,6 +84,133 @@ class BaseTable():
         self.filter_model: FilterBaseModel = None
         self.validate_filter_field_types = True
         self.df: ps.DataFrame = None
+        # For table-based class methods
+        self.cache_dir = eval.cache_dir
+        self.dataset_dir = eval.dataset_dir
+        self.scripts_dir = eval.scripts_dir
+        self.locations_cache_dir = Path(
+            self.cache_dir,
+            const.LOADING_CACHE_DIR,
+            const.LOCATIONS_DIR
+        )
+        self.crosswalk_cache_dir = Path(
+            self.cache_dir,
+            const.LOADING_CACHE_DIR,
+            const.LOCATION_CROSSWALKS_DIR
+        )
+        self.attributes_cache_dir = Path(
+            self.cache_dir,
+            const.LOADING_CACHE_DIR,
+            const.LOCATION_ATTRIBUTES_DIR
+        )
+        self.secondary_cache_dir = Path(
+            self.cache_dir,
+            const.LOADING_CACHE_DIR,
+            const.SECONDARY_TIMESERIES_DIR
+        )
+        self.primary_cache_dir = Path(
+            self.cache_dir,
+            const.LOADING_CACHE_DIR,
+            const.PRIMARY_TIMESERIES_DIR
+        )
+        # End table-based class methods
+
+    def _raise_missing_table_error():
+        """Raise an error if the table does not exist."""
+        err_msg = (
+            "The requested table does not exist in the dataset."
+            " Please load it first."
+        )
+        logger.error(err_msg)
+        raise ValueError(err_msg)
+
+    def _dir_is_emtpy(self, extension_wildcard: str = "*.parquet") -> bool:
+        """Check if the directory contains files of specified extension."""
+        return len(list(self.dir.glob(extension_wildcard))) == 0
+
+    def _read_spark_df(self):
+        """Read data from directory."""
+        self.df = (
+            self.spark.read.format("parquet")
+            .option("recursiveFileLookup", "true")
+            .option("mergeSchema", "true")
+            .option("header", True)
+            .load(str(self.dir))
+        )
+
+    def _load_timeseries(
+        self,
+        in_path: Union[Path, str],
+        cache_path: Path,
+        pattern="**/*.parquet",
+        timeseries_type: str = None,
+        field_mapping: dict = None,
+        constant_field_values: dict = None,
+        **kwargs
+    ):
+        """Import timeseries helper."""
+        convert_timeseries(
+            in_path=in_path,
+            out_path=cache_path,
+            field_mapping=field_mapping,
+            constant_field_values=constant_field_values,
+            pattern=pattern,
+            **kwargs
+        )
+
+        if pattern.endswith(".csv"):
+            pattern = pattern.replace(".csv", ".parquet")
+        elif pattern.endswith(".nc"):
+            pattern = pattern.replace(".nc", ".parquet")
+
+        validate_and_insert_timeseries(
+            in_path=cache_path,
+            dataset_path=self.dataset_dir,
+            timeseries_type=timeseries_type,
+            pattern=pattern
+        )
+
+        self._read_spark_df()
+
+        return self
+
+    def _load_locations_attributes(
+        self,
+        in_path: Union[Path, str],
+        pattern: str = None,
+        field_mapping: dict = None,
+        **kwargs
+    ):
+        """Load location attributes helper."""
+        convert_location_attributes(
+            in_path,
+            self.attributes_cache_dir,
+            pattern=pattern,
+            field_mapping=field_mapping,
+            **kwargs
+        )
+        validate_and_insert_location_attributes(
+            self.attributes_cache_dir, self.dataset_dir
+        )
+
+    def _load_location_crosswalks(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        pattern: str = None,
+        **kwargs
+    ):
+        """Load location crosswalks helper."""
+        convert_location_crosswalks(
+            in_path,
+            self.crosswalk_cache_dir,
+            field_mapping=field_mapping,
+            pattern=pattern,
+            **kwargs
+        )
+        validate_and_insert_location_crosswalks(
+            self.crosswalk_cache_dir, self.dataset_dir
+        )
 
     def query(
         self,
@@ -148,6 +303,9 @@ class BaseTable():
         >>> ]).to_pandas()
 
         """
+        logger.info("Performing the query.")
+        if self.df is None:
+            self._raise_missing_table_error()
         if filters is not None:
             self.df = validate_and_apply_filters(
                 sdf=self.df,
@@ -241,6 +399,8 @@ class BaseTable():
 
         """
         logger.info(f"Setting filter {filter}.")
+        if self.df is None:
+            self._raise_missing_table_error()
         self.df = validate_and_apply_filters(
             sdf=self.df,
             filters=filters,
@@ -279,6 +439,9 @@ class BaseTable():
         >>> ).to_pandas()
 
         """
+        logger.info(f"Setting order_by {fields}.")
+        if self.df is None:
+            self._raise_missing_table_error()
         self.df = order_df(self.df, fields)
         return self
 
@@ -292,6 +455,8 @@ class BaseTable():
 
     def to_pandas(self):
         """Return Pandas DataFrame."""
+        if self.df is None:
+            self._raise_missing_table_error()
         return self.df.toPandas()
 
     def to_geopandas(self):
@@ -319,6 +484,8 @@ class BaseTable():
         >>> )
         >>> ts_df.head()
         """
+        if self.df is None:
+            self._raise_missing_table_error()
         return self.df
 
 
@@ -331,14 +498,14 @@ class UnitTable(BaseTable):
         self.dir = eval.units_dir
         self.table_model = Unit
         self.filter_model = UnitFilter
-
-        self.df = (
-            self.spark.read.format("csv")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> UnitFields:
         """Get the unit fields enum."""
@@ -347,6 +514,28 @@ class UnitTable(BaseTable):
             "UnitFields",
             {field: field for field in fields_list}
         )
+
+    def add(
+        self,
+        unit: Union[Unit, List[Unit]]
+    ):
+        """Add a unit to the evaluation.
+
+        Parameters
+        ----------
+        unit : Union[Unit, List[Unit]]
+            The unit domain to add.
+
+        Example
+        -------
+        >>> from teehr.models.domain_tables import Unit
+        >>> unit = Unit(
+        >>>     name="m^3/s",
+        >>>     long_name="Cubic meters per second"
+        >>> )
+        >>> eval.load.add_unit(unit)
+        """
+        add_unit(self.dataset_dir, unit)
 
 
 class VariableTable(BaseTable):
@@ -358,14 +547,14 @@ class VariableTable(BaseTable):
         self.dir = eval.variables_dir
         self.table_model = Variable
         self.filter_model = VariableFilter
-
-        self.df = (
-            self.spark.read.format("csv")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> VariableFields:
         """Get the variable fields enum."""
@@ -374,6 +563,28 @@ class VariableTable(BaseTable):
             "VariableFields",
             {field: field for field in fields_list}
         )
+
+    def add(
+        self,
+        variable: Union[Variable, List[Variable]]
+    ):
+        """Add a unit to the evaluation.
+
+        Parameters
+        ----------
+        variable : Union[Variable, List[Variable]]
+            The variable domain to add.
+
+        Example
+        -------
+        >>> from teehr.models.domain_tables import Variable
+        >>> variable = Variable(
+        >>>     name="streamflow_hourly_inst",
+        >>>     long_name="Instantaneous streamflow"
+        >>> )
+        >>> eval.load.add_variable(variable)
+        """
+        add_variable(self.dataset_dir, variable)
 
 
 class AttributeTable(BaseTable):
@@ -385,14 +596,14 @@ class AttributeTable(BaseTable):
         self.dir = eval.attributes_dir
         self.table_model = Attribute
         self.filter_model = AttributeFilter
-
-        self.df = (
-            self.spark.read.format("csv")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> AttributeFields:
         """Get the attribute fields enum."""
@@ -401,6 +612,29 @@ class AttributeTable(BaseTable):
             "AttributeFields",
             {field: field for field in fields_list}
         )
+
+    def add_attribute(
+        self,
+        attribute: Union[Attribute, List[Attribute]]
+    ):
+        """Add an attribute to the evaluation.
+
+        Parameters
+        ----------
+        attribute : Union[Attribute, List[Attribute]]
+            The attribute domain to add.
+
+        Example
+        -------
+        >>> from teehr.models.domain_tables import Attribute
+        >>> attribute = Attribute(
+        >>>     name="drainage_area",
+        >>>     type="continuous",
+        >>>     description="Drainage area in square kilometers"
+        >>> )
+        >>> eval.load.add_attribute(attribute)
+        """
+        add_attribute(self.dataset_dir, attribute)
 
 
 class ConfigurationTable(BaseTable):
@@ -412,14 +646,14 @@ class ConfigurationTable(BaseTable):
         self.dir = eval.configurations_dir
         self.table_model = Configuration
         self.filter_model = ConfigurationFilter
-
-        self.df = (
-            self.spark.read.format("csv")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> ConfigurationFields:
         """Get the configuration fields enum."""
@@ -428,6 +662,30 @@ class ConfigurationTable(BaseTable):
             "ConfigurationFields",
             {field: field for field in fields_list}
         )
+
+    def add(
+        self,
+        configuration: Union[Configuration, List[Configuration]]
+    ):
+        """Add a configuration domain to the evaluation.
+
+        Parameters
+        ----------
+        configuration : Union[Configuration, List[Configuration]]
+            The configuration domain to add.
+
+        Example
+        -------
+        >>> from teehr.models.domain_tables import Configuration
+        >>> configuration = Configuration(
+        >>>     name="usgs_observations",
+        >>>     type="primary",
+        >>>     description="USGS observations",
+        >>> )
+        >>> eval.load.add_configuration(configuration)
+
+        """
+        add_configuration(self.dataset_dir, configuration)
 
 
 class LocationTable(BaseTable):
@@ -439,14 +697,14 @@ class LocationTable(BaseTable):
         self.dir = eval.locations_dir
         self.table_model = Location
         self.filter_model = LocationFilter
-
-        self.df = (
-            self.spark.read.format("parquet")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> LocationFields:
         """Get the location fields enum."""
@@ -458,7 +716,56 @@ class LocationTable(BaseTable):
 
     def to_geopandas(self):
         """Return GeoPandas DataFrame."""
+        if self.df is None:
+            err_msg = (
+                "No location data exists in the dataset."
+                " Please load it first using load_spatial()."
+            )
+            logger.error(err_msg)
         return df_to_gdf(self.to_pandas())
+
+    def load_spatial(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        pattern: str = "**/*.parquet",
+        **kwargs
+    ):
+        """Import geometry data.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            The input file or directory path.
+            Any file format that can be read by GeoPandas.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        pattern : str, optional (default: "**/*.parquet")
+            The pattern to match files.
+            Only used when in_path is a directory.
+        **kwargs
+            Additional keyword arguments are passed to GeoPandas read_file().
+
+        File is first read by GeoPandas, field names renamed and
+        then validated and inserted into the dataset.
+
+        The TEEHR Location Crosswalk table schema includes fields:
+        - id
+        - name
+        - geometry
+        """
+        convert_locations(
+            in_path,
+            self.locations_cache_dir,
+            field_mapping=field_mapping,
+            pattern=pattern,
+            **kwargs
+        )
+        validate_and_insert_locations(
+            self.locations_cache_dir,
+            self.dataset_dir
+        )
 
 
 class LocationAttributeTable(BaseTable):
@@ -470,14 +777,14 @@ class LocationAttributeTable(BaseTable):
         self.dir = eval.location_attributes_dir
         self.table_model = LocationAttribute
         self.filter_model = LocationAttributeFilter
-
-        self.df = (
-            self.spark.read.format("parquet")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> LocationAttributeFields:
         """Get the location attribute fields enum."""
@@ -489,7 +796,79 @@ class LocationAttributeTable(BaseTable):
 
     def to_geopandas(self):
         """Return GeoPandas DataFrame."""
+        if self.df is None:
+            err_msg = (
+                "No location attributes data exists in the dataset."
+                " Please load it first using load_parquet()"
+                " or load_csv()."
+            )
+            logger.error(err_msg)
+            raise ValueError(err_msg)
         return join_geometry(self.df, self.eval.locations.to_sdf())
+
+    def load_parquet(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        **kwargs
+    ):
+        """Import location_attributes from parquet file format.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            The input file or directory path.
+            Parquet file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        **kwargs
+            Additional keyword arguments are passed to pd.read_parquet().
+
+        The TEEHR Location Attribute table schema includes fields:
+        - location_id
+        - attribute_name
+        - value
+        """
+        pattern = "**/*.parquet"
+        self._load_locations_attributes(
+            in_path=in_path,
+            pattern=pattern,
+            field_mapping=field_mapping,
+            **kwargs
+        )
+
+    def load_csv(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        **kwargs
+    ):
+        """Import location_attributes from CSV file format.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            The input file or directory path.
+            CSV file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        **kwargs
+            Additional keyword arguments are passed to pd.read_parquet().
+
+        The TEEHR Location Attribute table schema includes fields:
+        - location_id
+        - attribute_name
+        - value
+        """
+        pattern = "**/*.csv"
+        self._load_locations_attributes(
+            in_path=in_path,
+            pattern=pattern,
+            field_mapping=field_mapping,
+            **kwargs
+        )
 
 
 class LocationCrosswalkTable(BaseTable):
@@ -501,14 +880,14 @@ class LocationCrosswalkTable(BaseTable):
         self.dir = eval.location_crosswalks_dir
         self.table_model = LocationCrosswalk
         self.filter_model = LocationCrosswalkFilter
-
-        self.df = (
-            self.spark.read.format("parquet")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> LocationCrosswalkFields:
         """Get the location crosswalk fields enum."""
@@ -520,9 +899,78 @@ class LocationCrosswalkTable(BaseTable):
 
     def to_geopandas(self):
         """Return GeoPandas DataFrame."""
+        if self.df is None:
+            err_msg = (
+                "No location crosswalk data exists in the dataset."
+                " Please load it first using load_parquet()."
+            )
+            logger.error(err_msg)
+            raise ValueError(err_msg)
         return join_geometry(
             self.df, self.eval.locations.to_sdf(),
             "primary_location_id"
+        )
+
+    def load_parquet(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        **kwargs
+    ):
+        """Import location crosswalks from parquet file format.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            The input file or directory path.
+            Parquet file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        **kwargs
+            Additional keyword arguments are passed to pd.read_csv()
+            or pd.read_parquet().
+
+        The TEEHR Location Crosswalk table schema includes fields:
+        - primary_location_id
+        - secondary_location_id
+        """
+        self._load_location_crosswalks(
+            in_path=in_path,
+            field_mapping=field_mapping,
+            pattern="**/*.parquet",
+            **kwargs
+        )
+
+    def load_csv(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        **kwargs
+    ):
+        """Import location crosswalks from CSV file format.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            The input file or directory path.
+            CSV file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        **kwargs
+            Additional keyword arguments are passed to pd.read_csv()
+            or pd.read_parquet().
+
+        The TEEHR Location Crosswalk table schema includes fields:
+        - primary_location_id
+        - secondary_location_id
+        """
+        self._load_location_crosswalks(
+            in_path=in_path,
+            field_mapping=field_mapping,
+            pattern="**/*.parquet",
+            **kwargs
         )
 
 
@@ -535,14 +983,14 @@ class PrimaryTimeseriesTable(BaseTable):
         self.dir = eval.primary_timeseries_dir
         self.table_model = Timeseries
         self.filter_model = TimeseriesFilter
-
-        self.df = (
-            self.spark.read.format("parquet")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> TimeseriesFields:
         """Get the timeseries fields enum."""
@@ -554,7 +1002,171 @@ class PrimaryTimeseriesTable(BaseTable):
 
     def to_geopandas(self):
         """Return GeoPandas DataFrame."""
+        if self.df is None:
+            err_msg = (
+                "No primary timeseries data exists in the dataset."
+                " Please load it first using either load_parquet(),"
+                " load_csv(), or load_netcdf()."
+            )
+            logger.error(err_msg)
+            raise ValueError(err_msg)
         return join_geometry(self.df, self.eval.locations.to_sdf())
+
+    def load_parquet(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        constant_field_values: dict = None,
+        **kwargs
+    ):
+        """Import primary timeseries parquet data.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            Path to the timeseries data (file or directory) in
+            parquet file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        constant_field_values : dict, optional
+            A dictionary mapping field names to constant values.
+            Format: {field_name: value}
+        **kwargs
+            Additional keyword arguments are passed to pd.read_parquet().
+
+        Includes validation and importing data to database.
+
+        The TEEHR Timeseries table schema includes fields:
+        - reference_time
+        - value_time
+        - configuration_name
+        - unit_name
+        - variable_name
+        - value
+        - location_id
+        """
+        logger.info(f"Loading primary timeseries parquet data: {in_path}")
+
+        pattern = "**/*.parquet"
+
+        self.primary_cache_dir.mkdir(parents=True, exist_ok=True)
+
+        self._load_timeseries(
+            in_path=in_path,
+            cache_path=self.primary_cache_dir,
+            pattern=pattern,
+            timeseries_type="primary",
+            field_mapping=field_mapping,
+            constant_field_values=constant_field_values,
+            **kwargs
+        )
+
+        return self
+
+    def load_csv(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        constant_field_values: dict = None,
+        **kwargs
+    ):
+        """Import primary timeseries csv data.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            Path to the timeseries data (file or directory) in
+            csv file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        constant_field_values : dict, optional
+            A dictionary mapping field names to constant values.
+            Format: {field_name: value}
+        **kwargs
+            Additional keyword arguments are passed to pd.read_csv().
+
+        Includes validation and importing data to database.
+
+        The TEEHR Timeseries table schema includes fields:
+        - reference_time
+        - value_time
+        - configuration_name
+        - unit_name
+        - variable_name
+        - value
+        - location_id
+        """
+        logger.info(f"Loading primary timeseries csv data: {in_path}")
+
+        pattern = "**/*.csv"
+
+        self.primary_cache_dir.mkdir(parents=True, exist_ok=True)
+
+        self._load_timeseries(
+            in_path=in_path,
+            cache_path=self.primary_cache_dir,
+            pattern=pattern,
+            timeseries_type="primary",
+            field_mapping=field_mapping,
+            constant_field_values=constant_field_values,
+            **kwargs
+        )
+
+        return self
+
+    def load_netcdf(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        constant_field_values: dict = None,
+        **kwargs
+    ):
+        """Import primary timeseries netcdf data.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            Path to the timeseries data (file or directory) in
+            netcdf file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        constant_field_values : dict, optional
+            A dictionary mapping field names to constant values.
+            Format: {field_name: value}
+        **kwargs
+            Additional keyword arguments are passed to xr.open_dataset().
+
+        Includes validation and importing data to database.
+
+        The TEEHR Timeseries table schema includes fields:
+        - reference_time
+        - value_time
+        - configuration_name
+        - unit_name
+        - variable_name
+        - value
+        - location_id
+        """
+        logger.info(f"Loading primary timeseries netcdf data: {in_path}")
+
+        pattern = "**/*.nc"
+
+        self.primary_cache_dir.mkdir(parents=True, exist_ok=True)
+
+        self._load_timeseries(
+            in_path=in_path,
+            cache_path=self.primary_cache_dir,
+            pattern=pattern,
+            timeseries_type="primary",
+            field_mapping=field_mapping,
+            constant_field_values=constant_field_values,
+            **kwargs
+        )
+
+        return self
 
 
 class SecondaryTimeseriesTable(BaseTable):
@@ -566,14 +1178,14 @@ class SecondaryTimeseriesTable(BaseTable):
         self.dir = eval.secondary_timeseries_dir
         self.table_model = Timeseries
         self.filter_model = TimeseriesFilter
-
-        self.df = (
-            self.spark.read.format("parquet")
-            .option("recursiveFileLookup", "true")
-            .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                .option("recursiveFileLookup", "true")
+                .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> TimeseriesFields:
         """Get the timeseries fields enum."""
@@ -585,7 +1197,169 @@ class SecondaryTimeseriesTable(BaseTable):
 
     def to_geopandas(self):
         """Return GeoPandas DataFrame."""
+        if self.df is None:
+            err_msg = (
+                "No secondary timeseries data exists in the dataset."
+                " Please load it first using either load_parquet(),"
+                " load_csv(), or load_netcdf()."
+            )
+            logger.error(err_msg)
+            raise ValueError(err_msg)
         return join_geometry(self.df, self.eval.locations.to_sdf())
+
+    def load_parquet(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        constant_field_values: dict = None,
+        **kwargs
+    ):
+        """Import secondary timeseries parquet data.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            Path to the timeseries data (file or directory) in
+            parquet file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        constant_field_values : dict, optional
+            A dictionary mapping field names to constant values.
+            Format: {field_name: value}
+        **kwargs
+            Additional keyword arguments are passed to pd.read_parquet().
+
+        Includes validation and importing data to database.
+
+        The TEEHR Timeseries table schema includes fields:
+        - reference_time
+        - value_time
+        - configuration_name
+        - unit_name
+        - variable_name
+        - value
+        - location_id
+        """
+        logger.info(f"Loading secondary timeseries parquet data: {in_path}")
+
+        pattern = "**/*.parquet"
+
+        self.secondary_cache_dir.mkdir(parents=True, exist_ok=True)
+
+        self._load_timeseries(
+            in_path=in_path,
+            cache_path=self.secondary_cache_dir,
+            pattern=pattern,
+            timeseries_type="secondary",
+            field_mapping=field_mapping,
+            constant_field_values=constant_field_values,
+            **kwargs
+        )
+
+        return self
+
+    def load_csv(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        constant_field_values: dict = None,
+        **kwargs
+    ):
+        """Import secondary timeseries csv data.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            Path to the timeseries data (file or directory) in
+            csv file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        constant_field_values : dict, optional
+            A dictionary mapping field names to constant values.
+            Format: {field_name: value}
+        **kwargs
+            Additional keyword arguments are passed to pd.read_csv().
+
+        Includes validation and importing data to database.
+
+        The TEEHR Timeseries table schema includes fields:
+        - reference_time
+        - value_time
+        - configuration_name
+        - unit_name
+        - variable_name
+        - value
+        - location_id
+        """
+        logger.info(f"Loading secondary timeseries csv data: {in_path}")
+
+        pattern = "**/*.csv"
+
+        self.secondary_cache_dir.mkdir(parents=True, exist_ok=True)
+
+        self._load_timeseries(
+            in_path=in_path,
+            cache_path=self.secondary_cache_dir,
+            pattern=pattern,
+            timeseries_type="secondary",
+            field_mapping=field_mapping,
+            constant_field_values=constant_field_values,
+            **kwargs
+        )
+
+        return self
+
+    def load_netcdf(
+        self,
+        in_path: Union[Path, str],
+        field_mapping: dict = None,
+        constant_field_values: dict = None,
+        **kwargs
+    ):
+        """Import secondary timeseries netcdf data.
+
+        Parameters
+        ----------
+        in_path : Union[Path, str]
+            Path to the timeseries data (file or directory) in
+            netcdf file format.
+        field_mapping : dict, optional
+            A dictionary mapping input fields to output fields.
+            Format: {input_field: output_field}
+        constant_field_values : dict, optional
+            A dictionary mapping field names to constant values.
+            Format: {field_name: value}
+        **kwargs
+            Additional keyword arguments are passed to xr.open_dataset().
+
+        Includes validation and importing data to database.
+
+        The TEEHR Timeseries table schema includes fields:
+        - reference_time
+        - value_time
+        - configuration_name
+        - unit_name
+        - variable_name
+        - value
+        - location_id
+        """
+        pattern = "**/*.nc"
+
+        self.secondary_cache_dir.mkdir(parents=True, exist_ok=True)
+
+        self._load_timeseries(
+            in_path=in_path,
+            cache_path=self.secondary_cache_dir,
+            pattern=pattern,
+            timeseries_type="secondary",
+            field_mapping=field_mapping,
+            constant_field_values=constant_field_values,
+            **kwargs
+        )
+
+        return self
 
 
 class JoinedTimeseriesTable(BaseTable):
@@ -598,14 +1372,14 @@ class JoinedTimeseriesTable(BaseTable):
         self.filter_model = JoinedTimeseriesFilter
         self.table_model = JoinedTimeseriesTable
         self.validate_filter_field_types = False
-
-        self.df = (
-            self.spark.read.format("parquet")
-            # .option("recursiveFileLookup", "true")
-            # .option("mergeSchema", "true")
-            .option("header", True)
-            .load(str(self.dir))
-        )
+        if not self._dir_is_emtpy():
+            self.df = (
+                self.spark.read.format("parquet")
+                # .option("recursiveFileLookup", "true")
+                # .option("mergeSchema", "true")
+                .option("header", True)
+                .load(str(self.dir))
+            )
 
     def field_enum(self) -> JoinedTimeseriesFields:
         """Get the joined timeseries fields enum."""
@@ -617,7 +1391,29 @@ class JoinedTimeseriesTable(BaseTable):
 
     def to_geopandas(self):
         """Return GeoPandas DataFrame."""
+        if self.df is None:
+            err_msg = (
+                "No joined timeseries data exists in the dataset."
+                " Please create it first using create()."
+            )
+            logger.error(err_msg)
+            raise ValueError(err_msg)
         return join_geometry(
             self.df, self.eval.locations.to_sdf(),
             "primary_location_id"
+        )
+
+    def create(self, execute_udf: bool = False):
+        """Create joined timeseries table.
+
+        Parameters
+        ----------
+        execute_udf : bool, optional
+            Execute UDFs, by default False
+        """
+        create_joined_timeseries_dataset(
+            self.spark,
+            self.dataset_dir,
+            self.scripts_dir,
+            execute_udf,
         )
